// hr-frontend/src/lib/api/generated/model/organizationDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { OrganizationCounts } from './organizationCounts';

export interface OrganizationDetail {
  id?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  organizationTypeId?: number;
  type?: string;
  typeCode?: string;
  status?: string;
  establishedDate?: string;
  registrationNumber?: string;
  taxNumber?: string;
  phone?: string;
  email?: string;
  fax?: string;
  website?: string;
  countryId?: number;
  country?: string;
  stateId?: number;
  state?: string;
  cityId?: number;
  city?: string;
  addressLine1?: string;
  addressLine2?: string;
  address?: string;
  postalCode?: string;
  logoDataUrl?: string;
  logoText?: string;
  summary?: OrganizationCounts;
}