// hr-frontend/src/lib/api/generated/model/updateOrganizationRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UpdateOrganizationRequest {
  /**
     * @minLength 0
     * @maxLength 100
     */
  code?: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  nameEn: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  nameAr: string;
  organizationTypeId: number;
  active?: boolean;
  establishedDate: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  registrationNumber: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  taxNumber: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  phone: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  email: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  fax?: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  website?: string;
  countryId: number;
  stateId: number;
  cityId: number;
  /**
     * @minLength 0
     * @maxLength 255
     */
  addressLine1: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  addressLine2?: string;
  /**
     * @minLength 0
     * @maxLength 20
     */
  postalCode?: string;
}